requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cssselect2==0.7.0
customtkinter==5.1.3
darkdetect==0.8.0
defusedxml==0.7.1
docopt==0.6.2
freetype-py==2.3.0
idna==3.4
Js2Py==0.74
lxml==4.9.2
packaging==23.1
Pillow==9.5.0
pipwin==0.5.2
pycairo==1.23.0
pycparser==2.21
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
reportlab==4.0.0
requests==2.30.0
rlPyCairo==0.2.0
six==1.16.0
soupsieve==2.4.1
svglib==1.5.1
tinycss2==1.2.1
tzdata==2023.3
tzlocal==5.0
urllib3==2.0.2
webencodings==0.5.1
wikipedia==1.4.0
Wikipedia-API==0.5.8
